requirements.txt
apache-airflow-providers-apache-kafka==1.6.1
graphviz==0.20.3
psycopg2-binary==2.9.9
kafka-python==2.0.2
apache-airflow-providers-common-sql==1.14.2
apache-airflow-providers-postgres==5.11.2
SQLAlchemy~=1.4.52
pendulum~=3.0.0
tabulate==0.9.0
python-dotenv
pika==1.3.2